requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
jax==0.4.10
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
pyside2==5.15.2.1